P4.2: single-source API contract types in src/types/api.ts
The delete-preview and admin-games response shapes were declared twice
(route + consuming page) and had already drifted on created_at
nullability. One declaration site now serves both sides; the admin
page's created_at sort is null-safe.
